requirements.txt
torch>=2.0.0
transformers>=4.38.0
datasets>=2.18.0
accelerate>=0.27.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
wandb>=0.16.0
sentencepiece>=0.1.99
protobuf>=3.20.0
pyyaml>=6.0.1
peft>=0.11.0
evaluate>=0.4.1
bert-score>=0.3.13
safetensors>=0.4.2
